+++ hapi-fhir-jpaserver-test-r4/src/test/java/ca/uhn/fhir/jpa/dao/r4/FhirResourceDaoR4CreateTest.java
@@ -125,6 +125,7 @@ public class FhirResourceDaoR4CreateTest extends BaseJpaR4Test {
 		runInTransaction(()->{
 			logAllTokenIndexes();
 			logAllStringIndexes();
+			logAllUriIndexes();
 			assertEquals(0, myResourceIndexedSearchParamStringDao.count());
 			assertEquals(0, myResourceIndexedSearchParamTokenDao.count());
 			assertEquals(0, myResourceIndexedSearchParamUriDao.count());
