+++ hapi-fhir-validation/src/test/java/org/hl7/fhir/dstu3/hapi/validation/DefaultProfileValidationSupportR5Test.java
@@ -16,8 +16,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class DefaultProfileValidationSupportR5Test extends BaseValidationTestWithInlineMocks {
 
-	private static FhirContext ourCtx = FhirContext.forR5Cached();
-	private DefaultProfileValidationSupport mySvc = new DefaultProfileValidationSupport(ourCtx);
+	private final FhirContext myCtx = FhirContext.forR5Cached();
+	private final DefaultProfileValidationSupport mySvc = new DefaultProfileValidationSupport(myCtx);
 
 	@Test
 	public void testNoDuplicates() {
@@ -36,7 +36,7 @@ public class DefaultProfileValidationSupportR5Test extends BaseValidationTestWit
 
 		// Verify
 		assertNotNull(allSps);
-		assertEquals(1263, allSps.size());
+		assertEquals(1243, allSps.size());
 	}
 
 }
